[webapp] create generic prompt component
