+++ webapp/src/components/CreateStagePrompt.vue
@@ -1,16 +1,17 @@
 <template lang="pug">
-.c-create-stage-prompt(@click="$emit('close')")
-	.prompt-wrapper(v-scrollbar.y="", @click.stop="")
-		.prompt-wrapper-inner
-			bunt-icon-button#btn-close(@click="$emit('close')") close
-			h1 {{ $t('CreateStagePrompt:headline:text') }}
-			form(@submit.prevent="create")
-				bunt-input(name="name", :label="$t('CreateStagePrompt:name:label')", icon="theater", :placeholder="$t('CreateStagePrompt:name:placeholder')")
-				bunt-button(type="submit") {{ $t('CreateStagePrompt:submit:label') }}
+prompt.c-create-stage-prompt(@close="$emit('close')")
+	.content
+		bunt-icon-button#btn-close(@click="$emit('close')") close
+		h1 {{ $t('CreateStagePrompt:headline:text') }}
+		form(@submit.prevent="create")
+			bunt-input(name="name", :label="$t('CreateStagePrompt:name:label')", icon="theater", :placeholder="$t('CreateStagePrompt:name:placeholder')")
+			bunt-button(type="submit") {{ $t('CreateStagePrompt:submit:label') }}
 </template>
 <script>
+import Prompt from 'components/Prompt'
+
 export default {
-	components: {},
+	components: { Prompt },
 	data () {
 		return {
 		}
@@ -30,41 +31,20 @@ export default {
 </script>
 <style lang="stylus">
 .c-create-stage-prompt
-	position: fixed
-	top: 0
-	left: 0
-	width: 100vw
-	height: var(--vh100)
-	z-index: 1000
-	background-color: $clr-secondary-text-light
-	display: flex
-	justify-content: center
-	align-items: center
-	.prompt-wrapper
-		card()
+	.content
 		display: flex
 		flex-direction: column
-		width: 480px
-		max-height: 80vh
-		.prompt-wrapper-inner
+		padding: 32px
+		position: relative
+		h1
+			margin: 0
+			text-align: center
+		p
+			max-width: 320px
+		form
 			display: flex
 			flex-direction: column
-			padding: 32px
-			position: relative
-			#btn-close
-				icon-button-style(style: clear)
-				position: absolute
-				top: 8px
-				right: 8px
-			h1
-				margin: 0
-				text-align: center
-			p
-				max-width: 320px
-			form
-				display: flex
-				flex-direction: column
-				align-self: stretch
-				.bunt-button
-					themed-button-primary()
+			align-self: stretch
+			.bunt-button
+				themed-button-primary()
 </style>
